backend/requirements.txt
fastapi
uvicorn 
langchain 
python-dotenv
langchain-community
pydantic
openai
sqlmodel
requests
sse-starlette
psycopg[binary]
asyncpg
uuid